modification des noms de new_main en main_RIAC. Ceci avant de faire le merge
